+++ Trail Composer/Trail Composer/Models/DTOs/TrailFromApi.cs
@@ -12,13 +12,13 @@ namespace Trail_Composer.Models.DTOs
 
         public string? Description { get; set; }
 
-        public decimal MinLongitude { get; set; }
+        public int MinLongitude { get; set; }
 
-        public decimal MaxLongitude { get; set; }
+        public int MaxLongitude { get; set; }
 
-        public decimal MinLatitude { get; set; }
+        public int MinLatitude { get; set; }
 
-        public decimal MaxLatitude { get; set; }
+        public int MaxLatitude { get; set; }
 
         public ICollection<int> CountryIds { get; set; } = new List<int>();
         public ICollection<int> PathTypeIds { get; set; } = new List<int>();
